+++ src/arch/pami/machine.C
@@ -102,7 +102,7 @@ std::atomic<int> ckExitComplete {0};
 
 #if CMK_ERROR_CHECKING
 static int checksum_flag = 0;
-unsigned char computeCheckSum(unsigned char *data, int len);
+unsigned char computeCheckSum(unsigned char *data, size_t len);
 
 #define CMI_SET_CHECKSUM(msg, len)      \
         if (checksum_flag)  {   \
